Split msg test file
